backend/requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
PyJWT==1.7.1
numpy==1.24.3
scikit-learn==1.2.2
pandas==2.0.2
jwt==1.3.1
python-decouple==3.8
python-jose==3.3.0
pymongo==4.3.3
requests==2.31.0 
bcrypt
